The part where the main loop of the processing sketch sits.
Github says that I have to place extra info in this extended description box, so I'm doing so. Honestly I would like to place all of the information in the short summary above, but I guess I have to endure with this form of torture and split my description in two different boxes. Are you still reading? well the I guess I should explain the purpose (porpoise) of this little sketch. Is a simple experiment, I'm kind of new to processing, so I want to experiment with it's features before I make something more complicated.